requirements.txt
langchain~=0.0.134
transformers~=4.27.4
sentence_transformers~=2.2.2
llama-index~=0.5.10
fastapi~=0.95.0
pydantic~=1.10.7
pinecone-client~=2.2.1